placing game logic into reducer
