+++ client/src/app/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import counterReducer from "../features/counter/counterSlice";
+import gameReducer from "../features/game/gameSlice";
 
 export default configureStore({
   reducer: {
-    counter: counterReducer,
+    game: gameReducer,
   },
 });
