+++ client/src/index.js
@@ -1,14 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import { App } from "./App";
 import store from "./app/store";
 import { Provider } from "react-redux";
+import { Game } from "./features/game/game";
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <Game />
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
